+++ src/Form/TiptapEditorType.php
@@ -73,9 +73,7 @@ final class TiptapEditorType extends AbstractType
         $resolver->setAllowedTypes('example', ['null', 'string', TiptapExample::class]);
         $resolver->setAllowedTypes('theme', ['string']);
 
-        $resolver->setNormalizer('theme', function (Options $options, string $value): string {
-            return $this->normalizeTheme($value);
-        });
+        $resolver->setNormalizer('theme', fn(Options $options, string $value): string => $this->normalizeTheme($value));
 
         $resolver->setNormalizer('example', static function (Options $options, mixed $value): ?TiptapExample {
             if ($value === null || $value === '') {